Compare type() output with class
This is more direct than checking against the type of a constant.  It
also fixes one potential bug where code behind a condition `type(...) == ()`
was unreachable.

These direct type comparisons are not a best practice in Python, usually
conditions will use duck-typing, but it makes sense as a way of handling the
data types included in our ground truth file.